+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-## master
+## 1.10.0
 
 * Fix creating a volume when server is specified by its name
 * Deprecate and ignore the `--window` flag on `hcloud server enable-backup`